Update/Fix documentation.
The documentation for the constructor seems to be outdated, so I updated it for the current constructor.

Also, when using Ctrl+Q, the docs for the IapConnector constructor shall show, but they don't because it's misplaced. It shall be on top of the whole class.

If a different documentation for the class and the constructor is desired, better put the constructor in the body of the class with its own doc, and add the documentation of the class on top of the class definition.